Refactoring: class GeoDataPointmap::BreakLineCancelDefinitionCommand moved to independent files.
