+++ libs/geodata/pointmap/geodatapointmap.cpp
@@ -14,6 +14,7 @@
 #include "private/geodatapointmap_addpointsetreferencecommand.h"
 #include "private/geodatapointmap_breaklineaddcommand.h"
 #include "private/geodatapointmap_breaklineaddpointcommand.h"
+#include "private/geodatapointmap_breaklinecanceldefinitioncommand.h"
 #include "private/geodatapointmap_breaklinefinishdefinitioncommand.h"
 #include "private/geodatapointmap_deletepointscommand.h"
 #include "private/geodatapointmap_editpointscommand.h"
@@ -75,42 +76,6 @@
 #include <vtkVertex.h>
 
 #include <iriclib_pointmap.h>
-
-class GeoDataPointmap::BreakLineCancelDefinitionCommand : public QUndoCommand
-{
-public:
-	BreakLineCancelDefinitionCommand(GeoDataPointmap* parent) :
-		QUndoCommand {GeoDataPointmap::tr("Cancel Defining Break Line")}
-	{
-		m_pointMap = parent;
-		m_breakLine = m_pointMap->m_activeBreakLine;
-	}
-	void redo() override {
-		m_pointMap->m_mouseEventMode = GeoDataPointmap::meNormal;
-		m_breakLine->setActive(false);
-		m_breakLine->setHidden(true);
-		m_pointMap->m_activeBreakLine = 0;
-		m_pointMap->m_breakLines.removeOne(m_breakLine);
-		m_pointMap->updateMouseCursor(m_pointMap->graphicsView());
-		m_pointMap->updateActionStatus();
-		m_pointMap->m_needRemeshing = true;
-		m_pointMap->setMapped(false);
-	}
-	void undo() override {
-		m_pointMap->m_mouseEventMode = GeoDataPointmap::meBreakLineAddNotPossible;
-		m_pointMap->m_activeBreakLine = m_breakLine;
-		m_breakLine->setActive(true);
-		m_breakLine->setHidden(false);
-		m_pointMap->m_breakLines.append(m_breakLine);
-		m_pointMap->updateMouseCursor(m_pointMap->graphicsView());
-		m_pointMap->updateActionStatus();
-		m_pointMap->m_needRemeshing = true;
-		m_pointMap->setMapped(false);
-	}
-private:
-	GeoDataPointmapBreakLine* m_breakLine;
-	GeoDataPointmap* m_pointMap;
-};
 
 const char* GeoDataPointmap::VALUES = "values";
