+++ libs/geodata/pointmap/private/geodatapointmap_breaklinecanceldefinitioncommand.h
@@ -1,0 +1,21 @@
+#ifndef GEODATAPOINTMAP_BREAKLINECANCELDEFINITIONCOMMAND_H
+#define GEODATAPOINTMAP_BREAKLINECANCELDEFINITIONCOMMAND_H
+
+#include "../geodatapointmap.h"
+
+#include <QUndoCommand>
+
+class GeoDataPointmap::BreakLineCancelDefinitionCommand : public QUndoCommand
+{
+public:
+	BreakLineCancelDefinitionCommand(GeoDataPointmap* parent);
+
+	void redo() override;
+	void undo() override;
+
+private:
+	GeoDataPointmapBreakLine* m_breakLine;
+	GeoDataPointmap* m_pointMap;
+};
+
+#endif // GEODATAPOINTMAP_BREAKLINECANCELDEFINITIONCOMMAND_H
